Move `?uptime` and `?about` to Stats cog.
